% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickview.R
\docType{package}
\name{quickview}
\alias{quickview}
\alias{quickview-package}
\title{Open data in a View tab or in external softwares}
\description{
RStudio addin to quickly inspect dataframes in a View tab or open them with default CSV/text editor softwares.
}
\keyword{internal}
