% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickview.R
\docType{package}
\name{quickview}
\alias{quickview}
\alias{quickview-package}
\title{Open datasets in a View tab}
\usage{
quickview()
}
\description{
RStudio addin to quickly inspect dataframes in a View tab.
}
\keyword{internal}
